starting to split things up
